requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-bigquery==3.13.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0